Fix sign-compare in array layer-count comparisons
Compare slice/layer indices against array_layer_count as size_t (the
uint16_t otherwise promotes to signed int, tripping -Wsign-compare under
-Werror across all host configs), and compare the descriptor's layer
count as int in the tests.

+++ engine/src/flutter/impeller/core/texture.cc
@@ -65,7 +65,7 @@ bool Texture::IsSliceValid(size_t slice) const {
     case TextureType::kTextureCube:
       return slice <= 5;
     case TextureType::kTexture2DArray:
-      return slice < desc_.array_layer_count;
+      return slice < static_cast<size_t>(desc_.array_layer_count);
   }
   FML_UNREACHABLE();
 }
